requirements.txt
mysql-connector-python~=9.4.0
uvicorn~=0.35.0
fastapi~=0.116.1